Fixes for code review
- Moving gcs utilities into the utilities for validations since it's more a validation method.
- Rename the file `orbax.py` to `test_config_util.py`
