+++ dags/orbax/maxtext_mtc_emergency_save_local.py
@@ -21,7 +21,8 @@ from dags.orbax.util import validation_util
 from dags.orbax.util import checkpoint_util
 from xlml.utils.xpk import BRANCH_ABHINAV_MTC
 from xlml.utils.gke import zone_to_region
-from dags.orbax.util import orbax
+from dags.orbax.util import test_config_util
+
 
 SCHEDULE = "0 13 * * *" if composer_env.is_prod_env() else None
 DAG_TEST_NAME = "maxtext_emc_and_mtc_orbax_save_local"
@@ -81,7 +82,7 @@ with models.DAG(
   # Only one set of test configurations (e.g., v5p-128) is supported at the moment.
   # Other configurations (e.g., v5e and/or v6e) may be introduced later.
   test_configs = [
-      orbax.TestConfig(
+      test_config_util.TestConfig(
           cluster=XpkClusters.TPU_V5P_128_CLUSTER,
           machine_type="ct5p-hightpu-4t",
           accelerator="v5p-128",
@@ -92,18 +93,18 @@ with models.DAG(
           step=100,
           checkpoint_step=200,
           local_checkpoint_step=20,
-          base_dir=orbax.DEFAULT_BUCKET,
+          base_dir=test_config_util.DEFAULT_BUCKET,
       ),
   ]
 
   task_groups = []
 
   for checkpointing in [
-      orbax.Checkpointing(
+      test_config_util.Checkpointing(
           name="mtc",  # Multi-tier Checkpointing
           enable_multi_tier_checkpointing=True,
       ),
-      orbax.Checkpointing(
+      test_config_util.Checkpointing(
           name="emc",  # Emergency Checkpointing
           enable_multi_tier_checkpointing=False,
       ),
@@ -130,7 +131,7 @@ with models.DAG(
             )
 
             workload_command = test_config.generate_workload_command(
-                checkpoint_dir=orbax.DEFAULT_RAM_DISK,
+                checkpoint_dir=test_config_util.DEFAULT_RAM_DISK,
                 run_name=run_name,
                 slice_num=slice_num,
                 out_folder=f"maxtext_{checkpointing.name}_orbax_save_local",
@@ -147,7 +148,7 @@ with models.DAG(
                 docker_image=image.value,
                 test_owner=test_owner.CAMILO_Q,
             ).run(
-                ramdisk_directory=orbax.DEFAULT_RAM_DISK,
+                ramdisk_directory=test_config_util.DEFAULT_RAM_DISK,
                 mtc_enabled=True,
                 xpk_branch=BRANCH_ABHINAV_MTC,
                 skip_post_process=True,
@@ -164,7 +165,7 @@ with models.DAG(
                     project_id=test_config.cluster.project,
                     location=zone_to_region(test_config.cluster.zone),
                     cluster_name=test_config.cluster.name,
-                    ram_disk=orbax.DEFAULT_RAM_DISK,
+                    ram_disk=test_config_util.DEFAULT_RAM_DISK,
                     start_time=start_time,
                     end_time=end_time,
                     steps_to_validate=steps_to_validate,
