+++ dags/orbax/maxtext_mtc_save_gcs.py
@@ -17,7 +17,7 @@ from dags.orbax.util import checkpoint_util
 from dags.orbax.util import validation_util
 from xlml.utils.xpk import BRANCH_ABHINAV_MTC
 from xlml.utils.gke import zone_to_region
-from dags.orbax.util import orbax
+from dags.orbax.util import test_config_util
 
 
 SCHEDULE = "0 15 * * *" if composer_env.is_prod_env() else None
@@ -69,11 +69,11 @@ with models.DAG(
     """,
     concurrency=2,
 ) as dag:
-  checkpointing = orbax.Checkpointing(
+  checkpointing = test_config_util.Checkpointing(
       name="mtc", enable_multi_tier_checkpointing=True
   )
   test_configs = [
-      orbax.TestConfig(
+      test_config_util.TestConfig(
           cluster=XpkClusters.TPU_V5P_128_CLUSTER,
           machine_type="ct5p-hightpu-4t",
           accelerator="v5p-128",
@@ -84,7 +84,7 @@ with models.DAG(
           step=200,
           checkpoint_step=300,
           local_checkpoint_step=20,
-          base_dir=orbax.DEFAULT_BUCKET,
+          base_dir=test_config_util.DEFAULT_BUCKET,
       ),
   ]
   for mode, image in DOCKER_IMAGES:
@@ -106,7 +106,7 @@ with models.DAG(
         )
 
         workload_command = test_config.generate_workload_command(
-            checkpoint_dir=orbax.DEFAULT_RAM_DISK,
+            checkpoint_dir=test_config_util.DEFAULT_RAM_DISK,
             run_name=run_name,
             slice_num=slice_num,
             out_folder=f"maxtext_mtc_orbax_save_gcs",
@@ -124,7 +124,7 @@ with models.DAG(
             docker_image=image.value,
             test_owner=test_owner.CAMILO_Q,
         ).run(
-            ramdisk_directory=orbax.DEFAULT_RAM_DISK,
+            ramdisk_directory=test_config_util.DEFAULT_RAM_DISK,
             mtc_enabled=True,
             xpk_branch=BRANCH_ABHINAV_MTC,
             skip_post_process=True,
@@ -138,7 +138,7 @@ with models.DAG(
             project_id=test_config.cluster.project,
             location=zone_to_region(test_config.cluster.zone),
             cluster_name=test_config.cluster.name,
-            ram_disk=orbax.DEFAULT_RAM_DISK,
+            ram_disk=test_config_util.DEFAULT_RAM_DISK,
             start_time=start_time,
             end_time=end_time,
             steps_to_validate=steps_to_validate,
@@ -154,7 +154,7 @@ with models.DAG(
             pod_pattern="multitier-driver",
             start_time=start_time,
             end_time=end_time,
-            checkpoint_dir=f"{orbax.DEFAULT_BUCKET}/{run_name}",
+            checkpoint_dir=f"{test_config_util.DEFAULT_BUCKET}/{run_name}",
         )
 
         # Final CPC cleanup to ensure symmetric start/end
