+++ dags/orbax/util/validation_util.py
@@ -1,15 +1,14 @@
 """Utilities to get workloads logs and some utils."""
 
 from datetime import datetime, timezone, timedelta
-from typing import Optional
+from typing import Optional, List
 from absl import logging
 import re
 
+from airflow.providers.google.cloud.operators.gcs import GCSHook
 from airflow.decorators import task
 from airflow.exceptions import AirflowFailException
 from google.cloud import logging as logging_api
-
-from dags.orbax.util import gcs
 
 
 @task
@@ -191,7 +190,7 @@ def validate_log_with_gcs(
           gcs_checkpoint_path = match_gcs.group(0)
           step = match_step.group(1)
           logging.info(f"get gcs path from: {gcs_checkpoint_path}")
-          bucket_files = gcs.get_gcs_checkpoint(
+          bucket_files = get_gcs_checkpoint(
               f"{checkpoint_dir}/{gcs_checkpoint_path}/"
           )
           logging.info(f"gcs bucket files lenght: {len(bucket_files)}")
@@ -220,10 +219,11 @@ def validate_log_with_gcs(
               f"Could not find gcs_checkpoint_path or step in line: {line}"
           )
 
-  # Compare last step found in replicator logs and last (only one)
-  # step extracted from filename bucket
+  # Extract the step number from the LAST RECORDED file in GCS bucket ended with .meta,
+  # which is prefixed with 's' and followed by digits
+  # eg. <name_job>-<cluster_config>-<date>-36-s60-n26-w0.meta
+  # The extracted step should be step=60 and is the last saved step.
   if len(gcs_save_step_list_bucket) > 0 and len(gcs_save_step_list) > 0:
-    # Extract s60 from  file name with extension .meta
     pattern_bucket_step = r"s(\d+)"
     raw_str_filename = gcs_save_step_list_bucket[-1]
     match = re.search(pattern_bucket_step, raw_str_filename)
@@ -313,3 +313,38 @@ def list_log_entries(
 
   logging.info(f"Log filter constructed: {log_filter}")
   return list(logging_client.list_entries(filter_=log_filter))
+
+
+def get_gcs_checkpoint(output_path: str) -> List[str]:
+  """
+  Lists files in a GCS bucket at a specified path.
+
+  This function uses the GCSHook to connect to Google Cloud Storage.
+  It parses the provided `output_path` to extract the bucket name and prefix,
+  and then lists all objects within that path.
+
+  Args:
+    output_path (str): The full gs:// path to the GCS bucket and prefix
+      (e.g., "gs://my-bucket/my-folder/").
+
+  Returns:
+    List[str]: A list of file names (keys) found in the specified GCS path.
+  """
+  hook = GCSHook()
+  pattern = re.compile(r"^gs://(?P<bucket>[^/]+)/(?P<prefix>.+)$")
+  m = pattern.match(output_path)
+
+  if not m:
+    logging.error(f"Invalid GCS path format: {output_path}")
+    return []
+
+  bucket_name = m.group("bucket")
+  prefix = m.group("prefix")
+
+  logging.info(f"output_path:{output_path}")
+  logging.info(f"bucket:{bucket_name}")
+  logging.info(f"prefix:{prefix}")
+
+  files = hook.list(bucket_name=bucket_name, prefix=prefix)
+  logging.info(f"Files ===> {files}")
+  return files
